+++ Constants.go
@@ -1,0 +1,6 @@
+package opspec
+
+const (
+  NameOfOpFile = "op.yml"
+  NameOfCollectionFile = "collection.yml"
+)