+++ compositionRoot.go
@@ -1,10 +1,11 @@
-package sdk
+package opspec
 
 //go:generate counterfeiter -o ./fakeCompositionRoot.go --fake-name fakeCompositionRoot ./ compositionRoot
 
 type compositionRoot interface {
   CreateOpUseCase() createOpUseCase
-  SetDescriptionOfOpUseCase() setDescriptionOfOpUseCase
+  SetCollectionDescriptionUseCase() setCollectionDescriptionUseCase
+  SetOpDescriptionUseCase() setOpDescriptionUseCase
 }
 
 func newCompositionRoot(
@@ -18,14 +19,20 @@ filesystem Filesystem,
     yamlCodec,
   )
 
-  setDescriptionOfOpUseCase := newSetDescriptionOfOpUseCase(
+  setCollectionDescriptionUseCase := newSetCollectionDescriptionUseCase(
+    filesystem,
+    yamlCodec,
+  )
+
+  setOpDescriptionUseCase := newSetOpDescriptionUseCase(
     filesystem,
     yamlCodec,
   )
 
   compositionRoot = &_compositionRoot{
     createOpUseCase:createOpUseCase,
-    setDescriptionOfOpUseCase: setDescriptionOfOpUseCase,
+    setCollectionDescriptionUseCase:setCollectionDescriptionUseCase,
+    setOpDescriptionUseCase: setOpDescriptionUseCase,
   }
 
   return
@@ -33,14 +40,19 @@ filesystem Filesystem,
 }
 
 type _compositionRoot struct {
-  createOpUseCase              createOpUseCase
-  setDescriptionOfOpUseCase setDescriptionOfOpUseCase
+  createOpUseCase                 createOpUseCase
+  setCollectionDescriptionUseCase setCollectionDescriptionUseCase
+  setOpDescriptionUseCase         setOpDescriptionUseCase
 }
 
 func (this _compositionRoot) CreateOpUseCase() createOpUseCase {
   return this.createOpUseCase
 }
 
-func (this _compositionRoot) SetDescriptionOfOpUseCase() setDescriptionOfOpUseCase {
-  return this.setDescriptionOfOpUseCase
+func (this _compositionRoot) SetCollectionDescriptionUseCase() setCollectionDescriptionUseCase {
+  return this.setCollectionDescriptionUseCase
+}
+
+func (this _compositionRoot) SetOpDescriptionUseCase() setOpDescriptionUseCase {
+  return this.setOpDescriptionUseCase
 }