+++ sdk_suite_test.go
@@ -1,4 +1,4 @@
-package sdk
+package opspec
 
 import (
   . "github.com/onsi/ginkgo"