+++ CHANGELOG.md
@@ -4,4 +4,5 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - create op use case
-- set description of op use case
+- set collection description use case
+- set op description use case